improve delete of shared media

+++ cdmlib-services/src/main/java/eu/etaxonomy/cdm/api/service/MediaServiceImpl.java
@@ -37,8 +37,10 @@ import eu.etaxonomy.cdm.model.media.Media;
 import eu.etaxonomy.cdm.model.media.MediaRepresentation;
 import eu.etaxonomy.cdm.model.media.Rights;
 import eu.etaxonomy.cdm.model.name.TaxonName;
+import eu.etaxonomy.cdm.model.occurrence.MediaSpecimen;
 import eu.etaxonomy.cdm.model.occurrence.SpecimenOrObservationBase;
 import eu.etaxonomy.cdm.model.taxon.Taxon;
+import eu.etaxonomy.cdm.model.taxon.TaxonBase;
 import eu.etaxonomy.cdm.persistence.dao.media.IMediaDao;
 import eu.etaxonomy.cdm.strategy.cache.common.IIdentifiableEntityCacheStrategy;
 
@@ -54,6 +56,13 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
 
 	 @Autowired
 	 private IDescriptionService descriptionService;
+
+	 @Autowired
+     private IOccurrenceService specimenService;
+	 @Autowired
+     private ITaxonService taxonService;
+	 @Autowired
+     private INameService nameService;
 
 	@Override
     public Pager<MediaKey> getMediaKeys(Set<Taxon> taxonomicScope, Set<NamedArea> geoScopes, Integer pageSize, Integer pageNumber, List<String> propertyPaths) {
@@ -102,7 +111,7 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
         Media media = this.load(mediaUuid);
 
         result = isDeletable(mediaUuid, config);
-
+        CdmBase updatedObject = null;
         if (result.isOk()){
             Set<CdmBase> references = commonService.getReferencingObjectsForDeletion(media);
             String message = null;
@@ -120,17 +129,18 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                                 continue;
                             } else if ( (config.isDeleteFromDescription() && config.getDeleteFrom() instanceof Taxon  && config.getDeleteFrom().getId() == desc.getTaxon().getId())|| config.isDeleteFromEveryWhere()){
                                 Taxon taxon = desc.getTaxon();
+                                updatedObject = taxon;
                                 desc.removeElement(textData);
                                 textData.removeMedia(media);
                                 if (desc.getElements().isEmpty()){
                                     taxon.removeDescription(desc);
                                 }
+
                             } else {
                                 // this should not be happen, because it is not deletable. see isDeletable
                                 result.setAbort();
-                                message = "The media could not be deleted because it is referenced by a taxon.";
-                                result.addException(new ReferencedObjectUndeletableException(message));
-                                return result;
+
+
                             }
                         } else if (description instanceof SpecimenDescription){
                             SpecimenDescription desc = HibernateProxyHelper.deproxy(description, SpecimenDescription.class);
@@ -138,6 +148,7 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                                 continue;
                             } else if ((config.isDeleteFromDescription() && config.getDeleteFrom() instanceof SpecimenOrObservationBase  && config.getDeleteFrom().getId() == desc.getDescribedSpecimenOrObservation().getId())  || config.isDeleteFromEveryWhere()){
                                 SpecimenOrObservationBase<?> specimen = desc.getDescribedSpecimenOrObservation();
+                                updatedObject = specimen;
                                 desc.removeElement(textData);
                                 textData.removeMedia(media);
                                 if (desc.getElements().isEmpty()){
@@ -146,16 +157,17 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                             } else {
                              // this should not be happen, because it is not deletable. see isDeletable
                                 result.setAbort();
-                                message = "The media could not be deleted because it is referenced by a specimen.";
-                                result.addException(new ReferencedObjectUndeletableException(message));
-                                return result;
+
+
                             }
                         }else if (description instanceof TaxonNameDescription){
                             TaxonNameDescription desc = HibernateProxyHelper.deproxy(description, TaxonNameDescription.class);
+
                             if (desc.getTaxonName() == null ){
                                 continue;
                             } else if ((config.isDeleteFromDescription() && config.getDeleteFrom() instanceof TaxonName  && config.getDeleteFrom().getId() == desc.getTaxonName().getId())   || config.isDeleteFromEveryWhere()){
                                 TaxonName name= desc.getTaxonName();
+                                updatedObject = name;
                                 desc.removeElement(textData);
                                 textData.removeMedia(media);
                                 if (desc.getElements().isEmpty()){
@@ -164,19 +176,40 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                             } else {
                              // this should not be happen, because it is not deletable. see isDeletable
                                 result.setAbort();
-                                message = "The media could not be deleted because it is referenced by a specimen.";
-                                result.addException(new ReferencedObjectUndeletableException(message));
-                                return result;
+
+
                             }
                         }
                     }
                 }else if (ref instanceof MediaRepresentation){
                     MediaRepresentation mediaRep = HibernateProxyHelper.deproxy(ref, MediaRepresentation.class);
                     media.removeRepresentation(mediaRep);
+                } else if (ref instanceof MediaSpecimen){
+                    MediaSpecimen mediaSpecimen = HibernateProxyHelper.deproxy(ref, MediaSpecimen.class);
+                    if (config.getDeleteFrom().getId() == mediaSpecimen.getId() || config.isDeleteFromEveryWhere()){
+                        mediaSpecimen.setMediaSpecimen(null);
+                        updatedObject = mediaSpecimen;
+                    }
                 }
 
             }
-            dao.delete(media);
+            if (result.isOk()){
+                dao.delete(media);
+            } else{
+                if (updatedObject instanceof
+                         TaxonBase){
+                    taxonService.update((TaxonBase)updatedObject);
+                    result.addUpdatedObject(updatedObject);
+                }
+                if (updatedObject instanceof
+                        TaxonName){
+                   nameService.update((TaxonName)updatedObject);
+               }
+                if (updatedObject instanceof
+                        SpecimenOrObservationBase){
+                   specimenService.update((SpecimenOrObservationBase)updatedObject);
+               }
+            }
         }
         return result;
     }
@@ -189,6 +222,10 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
         Set<CdmBase> references = commonService.getReferencingObjectsForDeletion(media);
         MediaDeletionConfigurator mediaConfig = (MediaDeletionConfigurator)config;
         CdmBase deleteFrom = mediaConfig.getDeleteFrom();
+
+        if (mediaConfig.isDeleteFromEveryWhere()){
+           return result;
+        }
         for (CdmBase ref: references){
             String message = null;
             if (ref instanceof MediaRepresentation){
@@ -203,7 +240,7 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                         if (desc.getTaxon() == null || (mediaConfig.isDeleteFromDescription() && (deleteFrom instanceof Taxon && ((Taxon)deleteFrom).getId() == desc.getTaxon().getId()))){
                             continue;
                         } else{
-                            message = "The media can't be deleted because it is referenced by a taxon. ("+desc.getTaxon().getTitleCache()+")";
+                            message = "The media can't be deleted from the database because it is referenced by another taxon. ("+desc.getTaxon().getTitleCache()+")";
                         }
 
                     } else if (description instanceof SpecimenDescription){
@@ -211,24 +248,24 @@ public class MediaServiceImpl extends IdentifiableServiceBase<Media,IMediaDao> i
                         if (desc.getDescribedSpecimenOrObservation() == null || (mediaConfig.isDeleteFromDescription() && (deleteFrom instanceof SpecimenOrObservationBase && ((SpecimenOrObservationBase)deleteFrom).getId() == desc.getDescribedSpecimenOrObservation().getId()))){
                             continue;
                         } else{
-                            message = "The media can't be deleted because it is referenced by a specimen or observation. ("+desc.getDescribedSpecimenOrObservation().getTitleCache()+")";
+                            message = "The media can't be deleted from the database because it is referenced by another specimen or observation. ("+desc.getDescribedSpecimenOrObservation().getTitleCache()+")";
                         }
                     } else if (description instanceof TaxonNameDescription){
                         TaxonNameDescription desc = HibernateProxyHelper.deproxy(description, TaxonNameDescription.class);
                         if (desc.getTaxonName() == null || (mediaConfig.isDeleteFromDescription() && (deleteFrom instanceof TaxonName && ((TaxonName)deleteFrom).getId() == desc.getTaxonName().getId()))){
                             continue;
                         } else{
-                            message = "The media can't be deleted because it is referenced by a specimen or observation. ("+desc.getDescribedSpecimenOrObservation().getTitleCache()+")";
+                            message = "The media can't be deleted from the database because it is referenced by another specimen or observation. ("+desc.getDescribedSpecimenOrObservation().getTitleCache()+")";
                         }
                     }
                 }
             } else {
-                message = "The media can't be deleted because it is referenced by a " + ref.getUserFriendlyTypeName() ;
+                message = "The media can't be completely deleted because it is referenced by another " + ref.getUserFriendlyTypeName() ;
             }
             if (message != null){
                 result.addException(new ReferencedObjectUndeletableException(message));
                 result.addRelatedObject(ref);
-                result.setAbort();
+
             }
 
         }
